added support for jupyter notebook in VSCode (#95)
* added support for jupyter notebook in VSCode

In VSCode you can open an .ipynb file and change the language of a
code cell. In this use case there are no file extensions used but only a
language ID is use. Added a document
selector which is based only on the SQL language ID.

* Added support for Jupyter Notebooks

VSCode can open and run Jupyter Notebooks files (.ipynb).

A Notebook cell often executes python code via an IPython kernel.

IPython kernel suppor magics which enable it to run other languages such
as SQL.

This PR adds support for the following magics: %%sql, %%sparksql, %%trino

It detects cells with these magics and changes the mime type of the notebook cells.

By setting a cell to mime type SQL it activates the sql-language-server
enabling code completion within a
Notebook.

Co-authored-by: cccs-jc <cccs-jc@cyber.gc.ca>